+++ common/core/desktop/tests/unit/kmnkbd/debug_api.cpp
@@ -13,175 +13,19 @@
 #include <keyman/keyboardprocessor.h>
 #include "path.hpp"
 #include "state.hpp"
-#include "../test_assert.h"
 #include "kmx/kmx_base.h"
-#include "kmx/kmx_xstring.h"
+
+#include "action_items.hpp"
+#include "debug_items.hpp"
+
+#include "../test_assert.h"
 
 using namespace km::kbp::kmx;
 
-namespace
+km_kbp_option_item test_env_opts[] =
 {
-
-const char *debug_item_types[] = {
-  "KM_KBP_DEBUG_BEGIN", // = 0,
-  "//KM_KBP_DEBUG_BEGIN_ANSI", // = 1, // not supported; instead rewrite ansi keyboards to Unicode with mcompile
-  "KM_KBP_DEBUG_GROUP_ENTER", // = 2,
-  "KM_KBP_DEBUG_GROUP_EXIT", // = 3,
-  "KM_KBP_DEBUG_RULE_ENTER", // = 4,
-  "KM_KBP_DEBUG_RULE_EXIT", // = 5,
-  "KM_KBP_DEBUG_MATCH_ENTER", // = 6,
-  "KM_KBP_DEBUG_MATCH_EXIT", // = 7,
-  "KM_KBP_DEBUG_NOMATCH_ENTER", // = 8,
-  "KM_KBP_DEBUG_NOMATCH_EXIT", // = 9,
-  "KM_KBP_DEBUG_END" // = 10,
+  KM_KBP_OPTIONS_END
 };
-
-void print_debug_item(const char *title, km_kbp_state_debug_item const & item) {
-  LPGROUP gp = static_cast<LPGROUP>(item.kmx_info.group);
-  LPKEY rule = static_cast<LPKEY>(item.kmx_info.rule);
-
-  std::cout
-    << "debug_item " << title << std::endl
-    << "  type:    " << debug_item_types[item.type] << std::endl
-    << "  flags:   " << item.flags << std::endl
-    << "  key_info: (vk: " << item.key_info.vk
-    <<    " mod:   " << item.key_info.modifier_state
-    <<    " char:  " << item.key_info.character << ")" << std::endl;
-
-  std::cout
-    << "  kmx_info: "<< std::endl
-    << "    context: '" << item.kmx_info.context << "'" << std::endl
-    << "    action:  " << item.kmx_info.first_action << std::endl;
-
-  if(gp) std::cout
-    << "    group:   " << gp->dpName << std::endl;
-  if(rule) std::cout
-    << "    rule: " << std::endl
-    << "      key:    " << rule->Key << std::endl
-    << "      shift:  " << rule->ShiftFlags << std::endl
-    << "      line:   " << rule->Line << std::endl;
-  std::cout
-    << "    offsets:  ";
-  std::copy(
-    std::begin(item.kmx_info.store_offsets),
-    std::end(item.kmx_info.store_offsets),
-    std::ostream_iterator<uint16_t>(std::cout, " ")
-  );
-  std::cout << std::endl;
-}
-
-bool are_store_offsets_equal(const uint16_t (&lhs)[DEBUG_STORE_OFFSETS_SIZE], const uint16_t (&rhs)[DEBUG_STORE_OFFSETS_SIZE]) {
-  return std::equal(std::begin(lhs), std::end(lhs), std::begin(rhs));
-  //  !memcmp(lhs, rhs, sizeof(lhs));
-}
-
-bool operator==(km_kbp_state_debug_item const & lhs,
-                km_kbp_state_debug_item const & rhs)
-{
-  auto result = (lhs.type == rhs.type && lhs.flags == rhs.flags && lhs.kmx_info.first_action == rhs.kmx_info.first_action);
-  if(result) {
-    LPGROUP lgp = static_cast<LPGROUP>(lhs.kmx_info.group), rgp = static_cast<LPGROUP>(rhs.kmx_info.group);
-    LPKEY lrule = static_cast<LPKEY>(lhs.kmx_info.rule), rrule = static_cast<LPKEY>(rhs.kmx_info.rule);
-    switch(lhs.type) {
-      case KM_KBP_DEBUG_BEGIN:
-        result = lhs.key_info.character == rhs.key_info.character &&
-                 lhs.key_info.modifier_state == rhs.key_info.modifier_state &&
-                 lhs.key_info.vk == rhs.key_info.vk;
-        break;
-      case KM_KBP_DEBUG_END:
-        break;
-      case KM_KBP_DEBUG_GROUP_ENTER:
-      case KM_KBP_DEBUG_GROUP_EXIT:
-      case KM_KBP_DEBUG_MATCH_ENTER:
-      case KM_KBP_DEBUG_MATCH_EXIT:
-      case KM_KBP_DEBUG_NOMATCH_ENTER:
-      case KM_KBP_DEBUG_NOMATCH_EXIT:
-        assert(lgp != nullptr);
-        assert(rgp != nullptr);
-        assert(lgp->dpName != nullptr);
-        assert(rgp->dpName != nullptr);
-        result = !u16cmp(lgp->dpName, rgp->dpName);
-        break;
-      case KM_KBP_DEBUG_RULE_ENTER:
-      case KM_KBP_DEBUG_RULE_EXIT:
-        assert(lgp != nullptr);
-        assert(rgp != nullptr);
-        assert(lgp->dpName != nullptr);
-        assert(rgp->dpName != nullptr);
-        assert(lrule != nullptr);
-        assert(rrule != nullptr);
-        result = !u16cmp(lgp->dpName, rgp->dpName) &&
-          lrule->Line == rrule->Line &&
-          lrule->Key == rrule->Key &&
-          lrule->ShiftFlags == rrule->ShiftFlags &&
-          u16cmp(lhs.kmx_info.context, rhs.kmx_info.context) == 0 &&
-          are_store_offsets_equal(lhs.kmx_info.store_offsets, rhs.kmx_info.store_offsets);
-        break;
-      default:
-        assert(false);
-        result = false;
-    }
-  }
-
-  if(!result) {
-    print_debug_item("actual", lhs);
-    print_debug_item("expected", rhs);
-  }
-  return result;
-}
-
-
-bool debug_items(km_kbp_state const * state,
-                  std::initializer_list<km_kbp_state_debug_item> const & expected)
-{
-  size_t n = 0;
-  auto act = km_kbp_state_debug_items(state, &n);
-
-  for (auto &rhs: expected) {
-    if(--n < 0) {
-      std::cout << "expected longer than actual" << std::endl;
-      return false;
-    }
-    if (!(*act++ == rhs)) return false;
-  }
-
-  if(n != 0) {
-    std::cout << "actual longer than expected" << std::endl;
-    return false;
-  }
-
-  return true;
-}
-
-  km_kbp_option_item test_env_opts[] =
-  {
-    KM_KBP_OPTIONS_END
-  };
-
-} // namespace
-
-
-// TODO: consider using same const approach in kmx_base.h
-//       but this ripples through a whole lot of code
-typedef struct tagDEBUG_GROUP
-{
-  KMX_WCHAR const * dpName;
-  KEY const * dpKeyArray;   // [LPKEY] address of first item in key array
-  KMX_WCHAR const * dpMatch;
-  KMX_WCHAR const * dpNoMatch;
-  KMX_DWORD cxKeyArray;   // in array entries
-  KMX_BOOL  fUsingKeys;   // group(xx) [using keys] <-- specified or not
-} DEBUG_GROUP, *LPDEBUG_GROUP;
-
-typedef struct tagDEBUG_KEY
-{
-  KMX_WCHAR Key;
-  KMX_DWORD Line;
-  KMX_DWORD ShiftFlags;
-  KMX_WCHAR const * dpOutput;
-  KMX_WCHAR const * dpContext;
-} DEBUG_KEY, *LPDEBUG_KEY;
-
 
 km_kbp_keyboard * test_kb = nullptr;
 km_kbp_state * test_state = nullptr;
@@ -233,6 +77,11 @@ void test_debugging_disabled() {
   assert(debug_items(test_state, {
     km_kbp_state_debug_item{KM_KBP_DEBUG_END}
   }));
+
+  assert(action_items(test_state, {
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('S')}},
+    {KM_KBP_IT_END}
+  }));
 }
 
 /**
@@ -249,9 +98,11 @@ void test_debugging_no_rule_match() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_EXIT, KM_KBP_DEBUG_FLAG_NOMATCH, {}, {u"", &gp, nullptr, {}, 1}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 1}}
   }));
-  // assert(action_items(test_state, {
-//TODO
-  // }));
+
+  assert(action_items(test_state, {
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('S')}},
+    {KM_KBP_IT_END}
+  }));
 }
 
 /**
@@ -267,6 +118,12 @@ void test_debugging_function_key() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_ENTER, 0, {}, {u"", &gp}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_EXIT, KM_KBP_DEBUG_FLAG_NOMATCH, {}, {u"", &gp, nullptr, {}, 1}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, KM_KBP_DEBUG_FLAG_OUTPUTKEYSTROKE, {}, {u"", nullptr, nullptr, {}, 1}}
+  }));
+
+  assert(action_items(test_state, {
+    {KM_KBP_IT_EMIT_KEYSTROKE},
+    {KM_KBP_IT_INVALIDATE_CONTEXT}, // It's a non character key that is not a modifier, so this is a hint that context may no longer be valid
+    {KM_KBP_IT_END}
   }));
 }
 
@@ -289,12 +146,22 @@ void test_basic_rule_matches() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 1}}, // action item will emit a default 'D'
   }));
 
+  assert(action_items(test_state, {
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('D')}},
+    {KM_KBP_IT_END}
+  }));
+
   try_status(km_kbp_process_event(test_state, KM_KBP_VKEY_E, KM_KBP_MODIFIER_SHIFT));
   assert(debug_items(test_state, {
     km_kbp_state_debug_item{KM_KBP_DEBUG_BEGIN, KM_KBP_DEBUG_FLAG_UNICODE, {KM_KBP_VKEY_E, KM_KBP_MODIFIER_SHIFT, 'E'}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_ENTER, 0, {}, {u"", &gp}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_EXIT, KM_KBP_DEBUG_FLAG_NOMATCH, {}, {u"", &gp, nullptr, {}, 1}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 1}}, // action item will emit a default 'E'
+  }));
+
+  assert(action_items(test_state, {
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('E')}},
+    {KM_KBP_IT_END}
   }));
 
   try_status(km_kbp_process_event(test_state, KM_KBP_VKEY_F, KM_KBP_MODIFIER_SHIFT));
@@ -306,6 +173,16 @@ void test_basic_rule_matches() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_EXIT, 0, {}, {u"", &gp, nullptr, {}, 5 /* from above */ }},
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 5 /* from above */ }},
   }));
+
+  assert(action_items(test_state, {
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv(u'\u0E04')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv(u'\u0E05')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv(u'\u0E06')}},
+    {KM_KBP_IT_END}
+  }));
+
 }
 
 /**
@@ -342,6 +219,13 @@ void test_multiple_groups() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 3}}, // action item will emit a 'b'
   }));
 
+  assert(action_items(test_state, {
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('a')}},
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('b')}},
+    {KM_KBP_IT_END}
+  }));
+
   try_status(km_kbp_process_event(test_state, KM_KBP_VKEY_2, 0));
   assert(debug_items(test_state, {
     km_kbp_state_debug_item{KM_KBP_DEBUG_BEGIN, KM_KBP_DEBUG_FLAG_UNICODE, {KM_KBP_VKEY_2, 0, '2'}},
@@ -363,6 +247,15 @@ void test_multiple_groups() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_EXIT, 0, {}, {u"", &gp, nullptr, {}, 4}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 4}}, // action item will "abc"
   }));
+
+  assert(action_items(test_state, {
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('a')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('b')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('c')}},
+    {KM_KBP_IT_END}
+  }));
+
 }
 
 /**
@@ -390,6 +283,14 @@ void test_store_offsets() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 4}}, // action item will emit a 'exay'
   }));
 
+  assert(action_items(test_state, {
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('e')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('x')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('a')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('y')}},
+    {KM_KBP_IT_END}
+  }));
+
   try_status(km_kbp_process_event(test_state, KM_KBP_VKEY_B, 0));
   assert(debug_items(test_state, {
     km_kbp_state_debug_item{KM_KBP_DEBUG_BEGIN, KM_KBP_DEBUG_FLAG_UNICODE, {KM_KBP_VKEY_B, 0, 'b'}},
@@ -404,21 +305,50 @@ void test_store_offsets() {
     km_kbp_state_debug_item{KM_KBP_DEBUG_GROUP_EXIT, 0, {}, {u"", &gp, nullptr, {}, 6}},
     km_kbp_state_debug_item{KM_KBP_DEBUG_END, 0, {}, {u"", nullptr, nullptr, {}, 6}},
   }));
+
+  assert(action_items(test_state, {
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_BACK},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('e')}},
+    {KM_KBP_IT_CHAR, {0,}, {km_kbp_usv('x')}},
+    {KM_KBP_IT_END}
+  }));
 }
 
 
 
 constexpr const auto help_str = "\
-debug_api [--color] <SOURCE_PATH>\n\
+debug_api [--color] <SOURCE_PATH>|--print-sizeof\n\
 \n\
-  --color      Force color output\n\
-  SOURCE_PATH  Path where debug_api.cpp is found; kmx files are\n\
-               located relative to this path.\n";
+  --color         Force color output\n\
+  --print-sizeof  Emit structure sizes for interop debug\n\
+  SOURCE_PATH     Path where debug_api.cpp is found; kmx files are\n\
+                  located relative to this path.\n";
 
 int error_args() {
   std::cerr << "debug_api: Invalid arguments." << std::endl;
   std::cout << help_str;
   return 1;
+}
+
+int print_sizeof() {
+  std::cout << std::endl;
+  std::cout << "keyboardprocessor.h:" << std::endl;
+  std::cout << "sizeof(km_kbp_context_item): " << sizeof(km_kbp_context_item) << std::endl;
+  std::cout << "sizeof(km_kbp_context_item): " << sizeof(km_kbp_context_item) << std::endl;
+  std::cout << "sizeof(km_kbp_action_item): " << sizeof(km_kbp_action_item) << std::endl;
+  std::cout << "sizeof(km_kbp_option_item): " << sizeof(km_kbp_option_item) << std::endl;
+  std::cout << "sizeof(km_kbp_keyboard_attrs): " << sizeof(km_kbp_keyboard_attrs) << std::endl;
+  std::cout << "sizeof(km_kbp_attr): " << sizeof(km_kbp_attr) << std::endl;
+  std::cout << std::endl;
+  std::cout << "keyboardprocessor_debug.h:" << std::endl;
+  std::cout << "sizeof(km_kbp_state_debug_item): " << sizeof(km_kbp_state_debug_item) << std::endl;
+  std::cout << "sizeof(km_kbp_state_debug_key_info): " << sizeof(km_kbp_state_debug_key_info) << std::endl;
+  std::cout << "sizeof(km_kbp_state_debug_kmx_info): " << sizeof(km_kbp_state_debug_kmx_info) << std::endl;
+  std::cout << "sizeof([enum] km_kbp_debug_type): " << sizeof(km_kbp_debug_type) << std::endl;
+  exit(1);
 }
 
 int main(int argc, char *argv []) {
@@ -428,6 +358,10 @@ int main(int argc, char *argv []) {
 
   if(argc < 2) {
     return error_args();
+  }
+
+  if(std::string(argv[1]) == "--print-sizeof") {
+    return print_sizeof();
   }
 
   auto arg_color = std::string(argv[1]) == "--color";
